base: Add virtual/cross-XXX recipe specific provider control handling
Currently, providers are set on a global config basis. This change allows
for a select set of providers configured in RECIPE_VIRTUAL_PROVIDERS to
be selected on a per recipe basis. This would allow for the selection of
virtual/cross-cc as gcc or clang for example.

Note that this only works for values in DEPENDS, values in task[depends]
flag values need to use the ${PREFERRED_PROVIDER_virtual/xxx} method, as
used in this patch in a handful of places.

The PROVIDERS are removed from the recipes so that if a version of the
dependency accidentally slips through, the build will fail and the user
can correct the issue.

(From OE-Core rev: c0d87f578ef1ca9ec69f807b1ab45453ae54d406)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>